컴포넌트 분리 , 컴포넌트 이름 변경
